btodb and dbtob need to work with 64-bit numbers (from karels)
SCCS-vsn: sys/hp300/include/param.h 8.2
SCCS-vsn: sys/i386/include/param.h 8.2
SCCS-vsn: sys/luna68k/include/param.h 8.2
SCCS-vsn: sys/pmax/include/param.h 8.2
SCCS-vsn: sys/news3400/include/param.h 8.2
SCCS-vsn: sys/sparc/include/param.h 8.2

+++ usr/src/sys/i386/include/param.h
@@ -7,7 +7,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)param.h	8.1 (Berkeley) %G%
+ *	@(#)param.h	8.2 (Berkeley) %G%
  */
 
 /*
@@ -94,9 +94,9 @@
 #define	btoc(x)	(((unsigned)(x)+(NBPG-1))>>PGSHIFT)
 
 #define	btodb(bytes)	 		/* calculates (bytes / DEV_BSIZE) */ \
-	((unsigned)(bytes) >> DEV_BSHIFT)
+	((bytes) >> DEV_BSHIFT)
 #define	dbtob(db)			/* calculates (db * DEV_BSIZE) */ \
-	((unsigned)(db) << DEV_BSHIFT)
+	((db) << DEV_BSHIFT)
 
 /*
  * Map a ``block device block'' to a file system block.
